lccde_project/pyproject.toml
[build-system]
requires = ['setuptools>=68', 'wheel']
build-backend = 'setuptools.build_meta'

[project]
name = 'lccde'
version = '0.1.0'
description = 'LCCDE pipeline extracted from notebook'
authors = [{name = 'Your Team'}]
requires-python = '>=3.10'
dependencies = ['pandas>=2.0','numpy>=1.26','scikit-learn>=1.4','joblib>=1.3','pyyaml>=6.0','matplotlib>=3.8']

[tool.setuptools.packages.find]
where = ['.']
include = ['lccde*']
